Updating logging_fat tests to work with OpenLiberty builds
