+++ dev/com.ibm.ws.logging_fat/.classpath
@@ -1,12 +1,18 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <classpath>
 	<classpathentry kind="src" path="fat/src"/>
+	<classpathentry kind="src" path="test-applications/TestApp/src"/>
+	<classpathentry kind="src" path="test-applications/broken-servlet/src"/>
+	<classpathentry kind="src" path="test-applications/ffdc-servlet/src"/>
+	<classpathentry kind="src" path="test-applications/logger-servlet/src"/>
+	<classpathentry kind="src" path="test-applications/missing-feature-servlet/src"/>
 	<classpathentry kind="src" path="test-bundles/message.source.handler/src"/>
 	<classpathentry kind="src" path="test-bundles/sample.source.handler/src"/>
 	<classpathentry kind="src" path="test-bundles/accesslog.source.handler/src"/>
 	<classpathentry kind="src" path="test-bundles/ffdc.source.handler/src"/>
 	<classpathentry kind="src" path="test-bundles/trace.source.handler/src"/>
-	<classpathentry kind="src" path="test-applications/TestApp/src"/>
+	<classpathentry kind="src" path="test-bundles/test.configuration.fallalloverthefloor.ibmfeature/src"/>
+	<classpathentry kind="src" path="test-bundles/test.configuration.fallalloverthefloor.userfeature/src"/>
 	<classpathentry kind="con" path="aQute.bnd.classpath.container"/>
 	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
 	<classpathentry kind="output" path="bin"/>
