--- dev/com.ibm.ws.logging_fat/build-test.xml
@@ -1,47 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-    Copyright (c) 2017 IBM Corporation and others.
-    All rights reserved. This program and the accompanying materials
-    are made available under the terms of the Eclipse Public License v1.0
-    which accompanies this distribution, and is available at
-    http://www.eclipse.org/legal/epl-v10.html
-   
-    Contributors:
-        IBM Corporation - initial API and implementation
- -->
-
-<project name="com.ibm.ws.logging_fat" default="buildandrun" basedir=".">
-	<mkdir dir="test-applications/broken-servlet/resources/WEB-INF" />
-	<mkdir dir="test-applications/logger-servlet/resources/WEB-INF" />
-	<mkdir dir="test-applications/missing-feature-servlet/resources/WEB-INF" />
-	<mkdir dir="test-applications/ffdc-servlet/resources/WEB-INF" />
-  <import file="../ant_build/public_imports/fat_imports.xml" />
-
-  <target name="addAutoInstallApplications"
-          description="Automatically Install applications on LibertyServer instance creation">
-    <!-- The bad config servers needs an app to configure weirdly, so give them the broken one -->
-    <mkdir dir="${basedir}/publish/servers/com.ibm.ws.logging.badconfig.user/apps" />
-    <mkdir dir="${basedir}/publish/servers/com.ibm.ws.logging.badconfig.ibm/apps" />
-
-    <copy file="${basedir}/publish/files/broken-servlet.zip"
-          tofile="${basedir}/publish/servers/com.ibm.ws.logging.badconfig.user/apps/broken-servlet.war" />
-    <copy file="${basedir}/publish/files/broken-servlet.zip"
-          tofile="${basedir}/publish/servers/com.ibm.ws.logging.badconfig.ibm/apps/broken-servlet.war" />
-
-    <mkdir dir="${basedir}/publish/servers/com.ibm.ws.logging.brokenserver/dropins" />
-    <move file="${basedir}/publish/files/broken-servlet.zip"
-          tofile="${basedir}/publish/servers/com.ibm.ws.logging.brokenserver/dropins/broken-servlet.war" />
- 
-    <mkdir dir="${basedir}/publish/servers/com.ibm.ws.logging.missingfeatureserver/dropins" />
-    <move file="${basedir}/publish/files/missing-feature-servlet.zip"
-          tofile="${basedir}/publish/servers/com.ibm.ws.logging.missingfeatureserver/dropins/missing-feature-servlet.war" />
-
-    <move file="${basedir}/publish/files/logger-servlet.zip"
-          tofile="${basedir}/publish/servers/com.ibm.ws.logging.healthcenter/dropins/logger-servlet.war"/>
-  	
-    <move file="${basedir}/publish/files/ffdc-servlet.zip"
-          tofile="${basedir}/publish/servers/com.ibm.ws.logging.isodateformat/dropins/ffdc-servlet.war"/>
- 
-  </target>
-
-</project>
