+++ dev/com.ibm.ws.logging_fat/fat/src/com/ibm/ws/logging/fat/StackTraceFilteringForBadlyWrittenThrowableTest.java
@@ -10,8 +10,11 @@
  *******************************************************************************/
 package com.ibm.ws.logging.fat;
 
+import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
+
+import com.ibm.websphere.simplicity.ShrinkHelper;
 
 import componenttest.topology.impl.LibertyServerFactory;
 
@@ -22,15 +25,19 @@ public class StackTraceFilteringForBadlyWrittenThrowableTest extends AbstractSta
     @BeforeClass
     public static void setUp() throws Exception {
         server = LibertyServerFactory.getLibertyServer("com.ibm.ws.logging.brokenserver");
-
         server.startServer();
-
-        //Make sure the application has come up before proceeding
-        server.addInstalledAppForValidation("broken-servlet");
+        ShrinkHelper.defaultDropinApp(server, "broken-servlet", "com.ibm.ws.logging.fat.broken.servlet");
 
         // Hit the servlet, to drive the error
         hitWebPage("broken-servlet", "BrokenWithABadlyWrittenThrowableServlet", true);
+    }
 
+    @AfterClass
+    public static void tearDown() throws Exception {
+        if (server != null && server.isStarted()) {
+            // The actual exception does not identify the class, SRVE0777E: Exception thrown by application class 'unknown.unknown:unknown'
+            server.stopServer(EXPECTED_EXCEPTION, "SRVE0777E: Exception thrown by application class");
+        }
     }
 
     @Test
@@ -48,13 +55,12 @@ public class StackTraceFilteringForBadlyWrittenThrowableTest extends AbstractSta
         assertMessagesLogDoesNotContain("The messages log had a trimmed stack trace in it.", INTERNAL_CLASSES_REGEXP);
         // We better have a line for the class that threw the exception
         assertMessagesLogContains("The console stack didn't show the originating class.",
-                                  "at com.ibm.ws.logging.fat.servlet.BrokenWithABadlyWrittenThrowableServlet.doGet");
+                                  "at com.ibm.ws.logging.fat.broken.servlet.BrokenWithABadlyWrittenThrowableServlet.doGet");
         // We also want at least one line about javax.servlet
         assertMessagesLogContains("The console stack was trimmed too aggressively.",
                                   "at javax.servlet.http.HttpServlet.service");
         assertMessagesLogContains("The console stack was apparently trimmed, but internal WAS classes got left in it",
-                                    "at com.ibm.ws.webcontainer");
-
+                                  "at com.ibm.ws.webcontainer");
     }
 
     @Test
