+++ dev/com.ibm.ws.logging_fat/fat/src/com/ibm/ws/logging/fat/StackTraceFilteringForLoggedExceptionParametersTest.java
@@ -10,8 +10,11 @@
  *******************************************************************************/
 package com.ibm.ws.logging.fat;
 
+import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
+
+import com.ibm.websphere.simplicity.ShrinkHelper;
 
 import componenttest.topology.impl.LibertyServerFactory;
 
@@ -21,17 +24,21 @@ public class StackTraceFilteringForLoggedExceptionParametersTest extends Abstrac
 
     @BeforeClass
     public static void setUp() throws Exception {
-        server = LibertyServerFactory
-                        .getLibertyServer("com.ibm.ws.logging.brokenserver");
+        server = LibertyServerFactory.getLibertyServer("com.ibm.ws.logging.brokenserver");
 
         server.startServer();
-
-        //Make sure the application has come up before proceeding
-        server.addInstalledAppForValidation("broken-servlet");
+        ShrinkHelper.defaultDropinApp(server, "broken-servlet", "com.ibm.ws.logging.fat.broken.servlet");
 
         // Hit the servlet, to drive the error
         hitWebPage("broken-servlet", "BrokenServlet", true);
 
+    }
+
+    @AfterClass
+    public static void tearDown() throws Exception {
+        if (server != null && server.isStarted()) {
+            server.stopServer(SPECIAL_BROKEN_EXCEPTION, "com.ibm.ws.logging.fat.broken.servlet.BrokenServlet.doGet");
+        }
     }
 
     @Test
@@ -42,7 +49,7 @@ public class StackTraceFilteringForLoggedExceptionParametersTest extends Abstrac
                                  INTERNAL_CLASSES_REGEXP);
         // We better have a line for the class that threw the exception
         assertConsoleLogContains("The console stack was trimmed too aggressively and stripped out our servlet.",
-                                 "at com.ibm.ws.logging.fat.servlet.BrokenServlet.doGet");
+                                 "at com.ibm.ws.logging.fat.broken.servlet.BrokenServlet.doGet");
         // We also want at least one line about javax.servlet
         assertConsoleLogContains("The console stack was was trimmed too aggressively and stripped out the API we're using",
                                  "at javax.servlet.http.HttpServlet.service");
