+++ dev/com.ibm.ws.logging_fat/fat/src/com/ibm/ws/logging/fat/StackTraceFilteringForLoggedExceptionWithACauseParametersTest.java
@@ -10,8 +10,11 @@
  *******************************************************************************/
 package com.ibm.ws.logging.fat;
 
+import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
+
+import com.ibm.websphere.simplicity.ShrinkHelper;
 
 import componenttest.topology.impl.LibertyServerFactory;
 
@@ -23,17 +26,20 @@ public class StackTraceFilteringForLoggedExceptionWithACauseParametersTest exten
 
     @BeforeClass
     public static void setUp() throws Exception {
-        server = LibertyServerFactory
-                        .getLibertyServer("com.ibm.ws.logging.brokenserver");
-
+        server = LibertyServerFactory.getLibertyServer("com.ibm.ws.logging.brokenserver");
         server.startServer();
-
-        //Make sure the application has come up before proceeding
-        server.addInstalledAppForValidation("broken-servlet");
+        ShrinkHelper.defaultDropinApp(server, "broken-servlet", "com.ibm.ws.logging.fat.broken.servlet");
 
         // Hit the servlet, to drive the error
         hitWebPage("broken-servlet", "BrokenWithACauseServlet", true);
 
+    }
+
+    @AfterClass
+    public static void tearDown() throws Exception {
+        if (server != null && server.isStarted()) {
+            server.stopServer(EXCEPTION_WITH_A_CAUSE, "com.ibm.ws.logging.fat.broken.servlet.BrokenWithACauseServlet.doGet");
+        }
     }
 
     @Test
@@ -44,7 +50,7 @@ public class StackTraceFilteringForLoggedExceptionWithACauseParametersTest exten
                                  INTERNAL_CLASSES_REGEXP);
         // We better have a line for the class that threw the exception
         assertConsoleLogContains("The console stack was trimmed too aggressively and stripped out our servlet.",
-                                 "at com.ibm.ws.logging.fat.servlet.BrokenWithACauseServlet.doGet");
+                                 "at com.ibm.ws.logging.fat.broken.servlet.BrokenWithACauseServlet.doGet");
 
         assertConsoleLogContains("The console log should say 'Caused by'", CAUSED_BY);
         assertConsoleLogContains("The console log include the root cause", CAUSE_EXCEPTION);
