+++ dev/com.ibm.ws.logging_fat/fat/src/com/ibm/ws/logging/fat/StackTraceFilteringForPrintedExceptionWithIBMCodeAtTopTest.java
@@ -10,8 +10,11 @@
  *******************************************************************************/
 package com.ibm.ws.logging.fat;
 
+import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
+
+import com.ibm.websphere.simplicity.ShrinkHelper;
 
 import componenttest.topology.impl.LibertyServerFactory;
 
@@ -22,15 +25,19 @@ public class StackTraceFilteringForPrintedExceptionWithIBMCodeAtTopTest extends 
     @BeforeClass
     public static void setUp() throws Exception {
         server = LibertyServerFactory.getLibertyServer("com.ibm.ws.logging.brokenserver");
-
         server.startServer();
-
-        //Make sure the application has come up before proceeding
-        server.addInstalledAppForValidation("broken-servlet");
+        ShrinkHelper.defaultDropinApp(server, "broken-servlet", "com.ibm.ws.logging.fat.broken.servlet");
 
         // Hit the servlet, to drive the error
         hitWebPage("broken-servlet", "IBMCodeAtTopExceptionPrintingServlet", false);
 
+    }
+
+    @AfterClass
+    public static void tearDown() throws Exception {
+        if (server != null && server.isStarted()) {
+            server.stopServer();
+        }
     }
 
     @Test
@@ -41,7 +48,7 @@ public class StackTraceFilteringForPrintedExceptionWithIBMCodeAtTopTest extends 
                                  INTERNAL_CLASSES_REGEXP);
         // We better have a line for the class that threw the exception
         assertConsoleLogContains("The console stack didn't show the originating class.",
-                                 "at com.ibm.ws.logging.fat.servlet.IBMCodeAtTopExceptionPrintingServlet.doGet");
+                                 "at com.ibm.ws.logging.fat.broken.servlet.IBMCodeAtTopExceptionPrintingServlet.doGet");
         // We also want at least one line about javax.servlet
         assertConsoleLogContains("The console stack was trimmed too aggressively.",
                                  "at javax.servlet.http.HttpServlet.service");
