+++ dev/com.ibm.ws.logging_fat/publish/servers/com.ibm.ws.logging.badconfig.ibm/server.xml
@@ -14,7 +14,7 @@
 
     <!-- Enable features -->
     <featureManager>
-        <feature>servlet-3.0</feature>
+        <feature>servlet-3.1</feature>
         <feature>usr:unconfigurableIbmFeature-1.0</feature>
     </featureManager>
