+++ dev/com.ibm.ws.logging_fat/publish/servers/com.ibm.ws.logging.brokenserver/server.xml
@@ -16,4 +16,7 @@
 	   <feature>jsp-2.2</feature>
     </featureManager>
     
+    <javaPermission className="java.util.PropertyPermission"  name="*" actions="read" />
+    
+    
 </server>
