+++ dev/com.ibm.ws.logging_fat/publish/servers/com.ibm.ws.logging.hidemessage/bootstrap.properties
@@ -9,4 +9,3 @@
 #     IBM Corporation - initial API and implementation
 ###############################################################################
 bootstrap.include=../testports.properties
-com.ibm.ws.logging.hideMessage=CWWKZ0058I
