+++ dev/com.ibm.ws.logging_fat/test-applications/broken-servlet/src/com/ibm/ws/logging/fat/broken/servlet/BrokenWithABadlyWrittenThrowableServlet.java
@@ -8,7 +8,7 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.logging.fat.servlet;
+package com.ibm.ws.logging.fat.broken.servlet;
 
 import java.io.IOException;
 
@@ -19,16 +19,16 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /**
- * Servlet which throws an exception.
+ * Servlet which throws an exception. The getStackTrace() method on the exception returns null.
  */
-@WebServlet("/BrokenWithACauseServlet")
-public class BrokenWithACauseServlet extends HttpServlet {
+@WebServlet("/BrokenWithABadlyWrittenThrowableServlet")
+public class BrokenWithABadlyWrittenThrowableServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
 
     /**
      * @see HttpServlet#HttpServlet()
      */
-    public BrokenWithACauseServlet() {
+    public BrokenWithABadlyWrittenThrowableServlet() {
         super();
     }
 
@@ -40,22 +40,20 @@ public class BrokenWithACauseServlet extends HttpServlet {
                          HttpServletResponse response) throws ServletException, IOException {
         response.setContentType("text/plain");
 
-        ReasonItAllWentWrongException reasonItAllWentWrongException = new ReasonItAllWentWrongException();
-        throw new BrokenWithACauseException(reasonItAllWentWrongException);
+        // Whoops, we seem to have a problem! Oh dear, how unexpected!
+        throw new BadlyWrittenException();
 
     }
 
-    static class BrokenWithACauseException extends RuntimeException {
-        public BrokenWithACauseException(ReasonItAllWentWrongException reasonItAllWentWrongException) {
-            super("arbitrary message", reasonItAllWentWrongException);
+    static class BadlyWrittenException extends RuntimeException {
+
+        private static final long serialVersionUID = 1L;
+
+        @Override
+        public StackTraceElement[] getStackTrace() {
+            // Can our logging code handle this?
+            return null;
         }
-
-        private static final long serialVersionUID = 1L;
-
-    }
-
-    static class ReasonItAllWentWrongException extends Exception {
-        private static final long serialVersionUID = 1L;
 
     }
