+++ dev/com.ibm.ws.logging_fat/test-applications/broken-servlet/src/com/ibm/ws/logging/fat/broken/servlet/BrokenWithACauseServlet.java
@@ -8,28 +8,27 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.logging.fat.servlet;
+package com.ibm.ws.logging.fat.broken.servlet;
 
 import java.io.IOException;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.Cookie;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /**
- * Servlet which prints an exception.
+ * Servlet which throws an exception.
  */
-@WebServlet("/SpecUsingServlet")
-public class SpecUsingServlet extends HttpServlet {
+@WebServlet("/BrokenWithACauseServlet")
+public class BrokenWithACauseServlet extends HttpServlet {
     private static final long serialVersionUID = 1L;
 
     /**
      * @see HttpServlet#HttpServlet()
      */
-    public SpecUsingServlet() {
+    public BrokenWithACauseServlet() {
         super();
     }
 
@@ -41,35 +40,23 @@ public class SpecUsingServlet extends HttpServlet {
                          HttpServletResponse response) throws ServletException, IOException {
         response.setContentType("text/plain");
 
-        response.getWriter().println("This servlet uses specification classes, but you shouldn't see this message.");
-        // Generate an exception which has a few lines of javax.servlet.* content in it
-        Cookie cookie = new Cookie(null, null);
-        response.addCookie(cookie);
+        ReasonItAllWentWrongException reasonItAllWentWrongException = new ReasonItAllWentWrongException();
+        throw new BrokenWithACauseException(reasonItAllWentWrongException);
 
     }
 
-    static class SpecialPrintingException extends Exception {
+    static class BrokenWithACauseException extends RuntimeException {
+        public BrokenWithACauseException(ReasonItAllWentWrongException reasonItAllWentWrongException) {
+            super("arbitrary message", reasonItAllWentWrongException);
+        }
 
         private static final long serialVersionUID = 1L;
 
     }
 
-    static class ExceptionGeneratingObject {
-        private final boolean shouldPrintException;
+    static class ReasonItAllWentWrongException extends Exception {
+        private static final long serialVersionUID = 1L;
 
-        public ExceptionGeneratingObject(boolean b) {
-            shouldPrintException = b;
-        }
-
-        @Override
-        public int hashCode() {
-            if (shouldPrintException) {
-                new SpecialPrintingException().printStackTrace();
-                return 1;
-            } else {
-                return -1;
-            }
-        }
     }
 
 }
