+++ dev/com.ibm.ws.logging_fat/test-applications/missing-feature-servlet/src/com/ibm/ws/logging/fat/missing/feature/servlet/MissingEntityManagerServlet.java
@@ -1,5 +1,5 @@
 /*******************************************************************************
- * Copyright (c) 2017 IBM Corporation and others.
+ * Copyright (c) 2013 IBM Corporation and others.
  * All rights reserved. This program and the accompanying materials
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
@@ -8,10 +8,14 @@
  * Contributors:
  *     IBM Corporation - initial API and implementation
  *******************************************************************************/
-package com.ibm.ws.logging.fat.servlet;
+package com.ibm.ws.logging.fat.missing.feature.servlet;
 
 import java.io.IOException;
+import java.util.List;
 
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+import javax.persistence.Query;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
@@ -19,16 +23,21 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 /**
- * Servlet which throws an exception.
+ * A servlet with a dependency on JPA, running in a server with no jpa feature loaded. What could
+ * possibly go wrong?
  */
-@WebServlet("/BrokenServlet")
-public class BrokenServlet extends HttpServlet {
+@WebServlet("/MissingEntityManagerServlet")
+public class MissingEntityManagerServlet extends HttpServlet {
+
     private static final long serialVersionUID = 1L;
+
+    @PersistenceContext(unitName = "thiswontworkpu")
+    private EntityManager em;
 
     /**
      * @see HttpServlet#HttpServlet()
      */
-    public BrokenServlet() {
+    public MissingEntityManagerServlet() {
         super();
     }
 
@@ -37,17 +46,13 @@ public class BrokenServlet extends HttpServlet {
      */
     @Override
     protected void doGet(HttpServletRequest request,
-                         HttpServletResponse response) throws ServletException, IOException {
-        response.setContentType("text/plain");
+                            HttpServletResponse response) throws ServletException, IOException {
+        response.getWriter().println("About to try and run a query on the entity manager we don't have.");
+        String query = "SELECT f FROM ChocolateOrder f";
+        Query q = em.createQuery(query);
 
-        // Whoops, we seem to have a problem! Oh dear, how unexpected!
-        throw new SpecialBrokenException();
-
-    }
-
-    static class SpecialBrokenException extends RuntimeException {
-
-        private static final long serialVersionUID = 1L;
+        List<?> list = q.getResultList();
+        response.getWriter().println("How did that work out?" + list);
 
     }
