+++ dev/com.ibm.ws.logging_fat/test-bundles/test.configuration.fallalloverthefloor.ibmfeature/bnd.bnd
@@ -11,7 +11,6 @@
 # Simulate a failure in Liberty code by having a user feature which looks like IBM code
 Bundle-Vendor: IBM
 bVersion=1.0.0
-Bundle-Version: 1.0.0
 Bundle-Name: Exception throwing feature
 Bundle-Description: This feature throws an exception
 Bundle-SymbolicName: test.configuration.fallalloverthefloor.ibmfeature; singleton:=true
